+++ synproxy/secret.c
@@ -85,10 +85,10 @@ static int verify_cookie46(
   struct conf *conf = synproxy->conf;
   int total_bits = 1 + conf->msslist_bits + conf->wscalelist_bits + 1;
   struct secret secret1;
-  uint16_t additional_bits = (isn>>(32-total_bits))&((1<<(total_bits-1))-1);
-  uint32_t bitmask = ((1<<(32-total_bits))-1);
-  uint32_t mssmask = ((1<<(conf->msslist_bits))-1);
-  uint32_t wsmask = ((1<<(conf->wscalelist_bits))-1);
+  uint32_t additional_bits = (isn>>(32-total_bits))&((1U<<(total_bits-1))-1U);
+  uint32_t bitmask = ((1U<<(32-total_bits))-1);
+  uint32_t mssmask = ((1U<<(conf->msslist_bits))-1);
+  uint32_t wsmask = ((1U<<(conf->wscalelist_bits))-1);
   struct siphash_ctx ctx;
   uint32_t hash;
   uint16_t *msstab = &DYNARR_GET(&conf->msslist, 0);
@@ -183,16 +183,16 @@ static uint32_t form_cookie46(
   uint32_t wsbits;
   uint32_t mssbits;
   uint32_t additional_bits;
-  int i;
-  int current_secret;
+  size_t i;
+  uint8_t current_secret;
   struct secret secret1;
   struct siphash_ctx ctx;
   uint32_t hash;
-  int wscnt = DYNARR_SIZE(&synproxy->conf->wscalelist);
-  int msscnt = DYNARR_SIZE(&synproxy->conf->msslist);
+  size_t wscnt = DYNARR_SIZE(&synproxy->conf->wscalelist);
+  size_t msscnt = DYNARR_SIZE(&synproxy->conf->msslist);
   uint16_t *msstab = &DYNARR_GET(&conf->msslist, 0);
   uint8_t *wstab = &DYNARR_GET(&conf->wscalelist, 0);
-  uint32_t bitmask = ((1<<(32-total_bits))-1);
+  uint32_t bitmask = ((1U<<(32-total_bits))-1U);
   for (i = 0; i < wscnt; i++)
   {
     if (wstab[i] > wscale)
@@ -216,7 +216,7 @@ static uint32_t form_cookie46(
   mssbits = i;
   sack_permitted = !!sack_permitted;
   additional_bits =
-      (sack_permitted<<(conf->wscalelist_bits+conf->msslist_bits))
+      (((uint32_t)sack_permitted)<<(conf->wscalelist_bits+conf->msslist_bits))
     | (mssbits<<conf->wscalelist_bits)
     | wsbits;
   if (pthread_rwlock_rdlock(&info->lock) != 0)
@@ -248,7 +248,7 @@ static uint32_t form_cookie46(
   siphash_feed_u64(&ctx, (((uint64_t)port1)<<48) | (((uint64_t)port2)<<32) | additional_bits);
   siphash_feed_u64(&ctx, other_isn);
   hash = siphash_get(&ctx) & bitmask;
-  return (current_secret<<31) | (additional_bits<<(32-total_bits)) | hash;
+  return (((uint32_t)current_secret)<<31) | (additional_bits<<(32-total_bits)) | hash;
 }
 
 uint32_t form_cookie(
@@ -290,13 +290,13 @@ static int verify_timestamp46(
   uint16_t *mss, uint8_t *wscale)
 {
   struct conf *conf = synproxy->conf;
-  int total_bits =
-    conf->ts_bits + conf->tsmsslist_bits + conf->tswscalelist_bits + 1;
+  uint32_t total_bits =
+    ((uint32_t)conf->ts_bits) + ((uint32_t)conf->tsmsslist_bits) + ((uint32_t)conf->tswscalelist_bits) + 1U;
   struct secret secret1;
-  uint16_t additional_bits = (isn>>(32-total_bits))&((1<<(total_bits-1))-1);
-  uint32_t bitmask = ((1<<(32-total_bits))-1);
-  uint32_t mssmask = ((1<<(conf->msslist_bits))-1);
-  uint32_t wsmask = ((1<<(conf->wscalelist_bits))-1);
+  uint16_t additional_bits = (isn>>(32-total_bits))&((1U<<(total_bits-1U))-1U);
+  uint32_t bitmask = ((1U<<(32-total_bits))-1U);
+  uint32_t mssmask = ((1U<<(conf->msslist_bits))-1U);
+  uint32_t wsmask = ((1U<<(conf->wscalelist_bits))-1U);
   struct siphash_ctx ctx;
   uint32_t hash;
   uint16_t *msstab = &DYNARR_GET(&conf->tsmsslist, 0);
@@ -332,11 +332,11 @@ static int verify_timestamp46(
   {
     if (wscale)
     {
-      *wscale = wstab[additional_bits&wsmask];
+      *wscale = wstab[((uint32_t)additional_bits)&wsmask];
     }
     if (mss)
     {
-      *mss = msstab[(additional_bits>>(conf->wscalelist_bits))&mssmask];
+      *mss = msstab[(((uint32_t)additional_bits)>>(conf->wscalelist_bits))&mssmask];
     }
     return 1;
   }
@@ -381,17 +381,17 @@ static uint32_t form_timestamp46(
   uint32_t wsbits;
   uint32_t mssbits;
   uint32_t additional_bits;
-  int i;
-  int current_secret;
+  size_t i;
+  uint8_t current_secret;
   struct secret secret1;
   struct siphash_ctx ctx;
   uint32_t hash;
-  int wscnt = DYNARR_SIZE(&synproxy->conf->tswscalelist);
-  int msscnt = DYNARR_SIZE(&synproxy->conf->tsmsslist);
+  size_t wscnt = DYNARR_SIZE(&synproxy->conf->tswscalelist);
+  size_t msscnt = DYNARR_SIZE(&synproxy->conf->tsmsslist);
   uint16_t *msstab = &DYNARR_GET(&conf->tsmsslist, 0);
   uint8_t *wstab = &DYNARR_GET(&conf->tswscalelist, 0);
-  uint32_t bitmask = ((1<<(32-total_bits))-1);
-  uint32_t ts = (gettime64() % 32000000)*(1<<conf->ts_bits) / 32000000;
+  uint32_t bitmask = ((1U<<(32-total_bits))-1U);
+  uint32_t ts = (gettime64() % 32000000)*(1U<<conf->ts_bits) / 32000000;
   for (i = 0; i < wscnt; i++)
   {
     if (wstab[i] > wscale)
@@ -445,7 +445,7 @@ static uint32_t form_timestamp46(
   }
   siphash_feed_u64(&ctx, (((uint64_t)port1)<<48) | (((uint64_t)port2)<<32) | additional_bits);
   hash = siphash_get(&ctx) & bitmask;
-  return (current_secret<<31) | (additional_bits<<(32-total_bits)) | hash;
+  return (((uint32_t)current_secret)<<31) | (additional_bits<<(32-total_bits)) | hash;
 }
 
 uint32_t form_timestamp(
